Nightly

+++ libraries/jomres/classes/jomres_install.class.php
@@ -4,7 +4,7 @@
  *
  * @author Vince Wooll <sales@jomres.net>
  *
- * @version Jomres 9.23.1
+ * @version Jomres 9.23.2
  *
  * @copyright	2005-2020 Vince Wooll
  * Jomres (tm) PHP, CSS & Javascript files are released under both MIT and GPL2 licenses. This means that you can choose the license that best suits your project, and use it accordingly
@@ -773,7 +773,10 @@ if (!defined('JOMRES_ROOT_DIRECTORY')) {
 	private function deleteFiles()
 	{
 		//delete jomres dir
-		$this->filesystem->deleteDir('local://'.JOMRES_ROOT_DIRECTORY);
+		if (is_dir(JOMRES_ROOT_DIRECTORY)) {
+			$this->filesystem->deleteDir('local://'.JOMRES_ROOT_DIRECTORY);
+		}
+
 		
 		//delete jomres_root file
 		$this->filesystem->delete('local://jomres_root.php');
